fix: Verify and fix misconceptions for students 51-60 (200150-200159)
Fixed mismatches:
- Flores_Luke_200151 Q4: Added INPUT001 (commented out import)
- Walker_Henry_200153 Q3: Added INPUT001 (commented out import)
- Jones_Sofia_200157 Q4: Added INPUT003 (not reading y3)
- Torres_Aiden_200159 Q2: Added OTHER001 (wrong formula)

+++ authentic_seeded/Flores_Luke_200151_Mixed/Q2.java
@@ -3,17 +3,17 @@
 // Assignment 2
 
 import java.util.Scanner;
-public class Q2
-{
+
+public class Q2 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		System.out.print("Enter the driving distance in miles: ");
-		double distance=sc.nextDouble();
+		double distance = sc.nextInt();
 		System.out.print("Enter miles per gallon: ");
-		double milesPerGallon=sc.nextDouble();
+		double milesPerGallon = sc.nextInt();
 		System.out.print("Enter price in $ per gallon: ");
-		double price=sc.nextDouble();
-		double cost=(distance/milesPerGallon)*price;
+		double price = sc.nextInt();
+		double cost = (distance / milesPerGallon) * price;
 		System.out.println("The cost of driving is $" + cost);
 	}
 }
