+++ authentic_seeded/Jones_Sofia_200157_Mixed/Q1.java
@@ -1,15 +1,13 @@
 import java.util.Scanner;
 
-public class Q1
-{
-    public static void main(String[] args)
-    {
+public class Q1 {
+    public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("Enter v0, v1, and t: ");
         double v0 = scanner.nextDouble();
         double v1 = scanner.nextDouble();
         double t = scanner.nextDouble();
-        double a = (v1 - v0) / t;
+        double a = (int) (v1 - v0) / (int) t;
         System.out.println("The average acceleration is " + a);
     }
 }
