+++ authentic_seeded/Lopez_Samuel_200154_DT001/Q4.java
@@ -3,11 +3,11 @@
 // Assignment 2
 
 import java.util.Scanner;
+
 public class Q4
 // TODO: Clean up code before submission
 {
-  public static void main(String[] args)
-  {
+  public static void main(String[] args) {
     Scanner in = new Scanner(System.in);
     System.out.println("Enter three points for a triangle.");
     System.out.print("(x1, y1): ");
@@ -22,7 +22,7 @@ public class Q4
     double side1 = Math.sqrt(Math.pow(x2 - x1, 2) + Math.pow(y2 - y1, 2));
     double side2 = Math.sqrt(Math.pow(x3 - x2, 2) + Math.pow(y3 - y2, 2));
     double side3 = Math.sqrt(Math.pow(x1 - x3, 2) + Math.pow(y1 - y3, 2));
-    double s = (side1 + side2 + side3) / 2;
+    double s = (int) (side1 + side2 + side3) / 2;
     double area = Math.sqrt(s * (s - side1) * (s - side2) * (s - side3));
     System.out.println("The area of the triangle is " + area);
   }
